% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Beazley}
\alias{Beazley}
\title{Beazley (sample of 1000)}
\format{
A data frame with 1000 rows and 4 variables
}
\source{
https://www.beazley.ox.ac.uk/pottery/default.htm
}
\usage{
data(Beazley)
}
\description{
A test dataset containing a data.frame how it should ideally be arranged to work with datplot.
Data are gathered from the Beazley Archive Pottery Database (BAPD) -- https://www.beazley.ox.ac.uk/pottery/default.htm
and transformed to work with datplot
}
\details{
\itemize{
  \item Identifier (Vase.Number in BAPD)
  \item Technique: Sample contains only red- or blackfigured objects
  \item DAT_min. Integer: lower range of the dating, BCE in negative numbers
  \item DAT_max. Integer: uppder range of the dating, BCE in negative numbers
}
}
\keyword{datasets}
