% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{calculate.outputrows}
\alias{calculate.outputrows}
\title{Calculate output rows}
\usage{
calculate.outputrows(DAT_mat, stepsize)
}
\arguments{
\item{DAT_mat}{todo}

\item{stepsize}{todo}
}
\value{
outputrows
}
\description{
approximation :(
}
