% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{check.number}
\alias{check.number}
\title{Check for numbers (internal)}
\usage{
check.number(value)
}
\arguments{
\item{value}{A value to check}
}
\value{
TRUE if value is any kind of number, FALSE if value is not
}
\description{
Checks if value is either numeric, integer or double
and and returns TRUE.
}
