% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{check.structure}
\alias{check.structure}
\title{Check if the structure is compatible with [datsteps()]  (internal)}
\usage{
check.structure(DAT_df, verbose = FALSE)
}
\arguments{
\item{DAT_df}{An object to check}

\item{verbose}{TRUE / FALSE: Should the function issue additional
messages pointing to possible inconsistencies and notify of methods?}
}
\value{
TRUE if object can be processed by [datsteps()], error / FALSE if not
}
\description{
Checks if the object passed to [datsteps()] can be used for
processing.
}
\keyword{internal}
