% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAT_df.R
\docType{data}
\name{DAT_df}
\alias{DAT_df}
\title{datplot Testing data}
\format{
A data frame with 10000 rows and 4 variables
}
\usage{
data(DAT_df)
}
\description{
A test dataset containing a data.frame how it should ideally be arranged to work with datplot.
Data at not real and illustrate some common problems such as lower and upper dating in the wrong columns.
}
\details{
\itemize{
  \item ID. Identifier of the Objects (has to be unique)
  \item carat. Grouping variable, such as a Type or a Findspot
  \item DAT_min. Integer: lower range of the dating, BCE in negative numbers
  \item DAT_max. Integer: uppder range of the dating, BCE in negative numbers
}
}
\keyword{datasets}
