% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot-package.R
\docType{package}
\name{datplot-package}
\alias{datplot}
\alias{datplot-package}
\title{datplot: Preparation of Object Dating Ranges for Density Plots (Aoristic Analysis)}
\description{
Converting date ranges into dating 'steps' eases the visualization of changes in e.g. pottery consumption, style and other variables over time. This package provides tools to process and prepare data for visualization and employs the concept of aoristic analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lsteinmann/datplot}
  \item \url{https://lsteinmann.github.io/datplot/}
  \item Report bugs at \url{https://github.com/lsteinmann/datplot/issues}
}

}
\author{
\strong{Maintainer}: Lisa Steinmann \email{lisa.steinmann@rub.de} (\href{https://orcid.org/0000-0002-2215-1243}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Barbora Weissova \email{barbora.weissova@rub.de} (\href{https://orcid.org/0000-0002-3297-6855}{ORCID}) [contributor]
}

}
\keyword{internal}
