% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_funs.R
\name{datsteps}
\alias{datsteps}
\title{Create 'steps' of dates for each object in a dataframe}
\usage{
datsteps(df, stepsize = 5)
}
\arguments{
\item{df}{a dataframe with 4 variable: ID, group, minimum date (num) maximum date (num), _must_ be in this order, colnames are irrelevant; each objects _must_ be one row.}

\item{stepsize}{defaults to 5. Number of years that should be considered a timestap/datestep.}
}
\value{
a larger dataframe with a number of steps for each object as well as a 'weight' value, that is a quantification of how well the object is dated (lesser values means objects are dated to larger timespans)
}
\description{
Requires a dataframe with 4 variable: one ID (ideally factor), one group (ideally factor),
a minimum date (int/numeric) and a maximum date (int/numeric). It's expected that dates BCE are
displayed as negative values while dates CE are positive. Ignoring this will cause problems when
crossing 0.
}
