% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_funs.R
\name{generate.stepsize}
\alias{generate.stepsize}
\title{Generate stepsize}
\usage{
generate.stepsize(timespans)
}
\arguments{
\item{timespans}{todo}
}
\value{
stepsize
}
\description{
todo
}
