% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.histogramscale.R
\name{get.histogramscale}
\alias{get.histogramscale}
\title{Scaling Factor for Combined Histogramm Plots}
\usage{
get.histogramscale(DAT_df, bw = "auto")
}
\arguments{
\item{DAT_df}{a dataframe as returned by datsteps}

\item{bw}{the bandwidth to use for the density function and histogram. Should be stepsize used to
create the dataframe. If a df as returned by datsteps() is given, stepsize is automatically assigned
using the corresponding attribute.}
}
\value{
the value with which to scale the density curve to a histogram plot so that both will be visible
}
\description{
Requires a dataframe as produced by datsteps().
Meaning 6 columns in the following order:
* ID,
* group,
* minimum/earliest date,
* maximum/latest date,
* weight,
* 'DAT_Steps'
}
\examples{
DAT_df_steps <- datsteps(DAT_df[1:100,], stepsize = 25)
get.histogramscale(DAT_df_steps)

}
