% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{get.probability}
\alias{get.probability}
\title{Calculate the probability for each year and each dated object}
\usage{
get.probability(DAT_min, DAT_max)
}
\arguments{
\item{DAT_min}{a vector containing the minimum date (a number) of each object}

\item{DAT_max}{a vector containing the maximum date (a number) of each object}
}
\value{
the probability for each object being dated to any single year
within the timespan (lesser value means object is dated to larger timespans,
i.e. with less confidence).
}
\description{
Calculates the probability of each object being dated into
each year / timeslot from two vectors of minimum and maximum
dating. Returns a vector of probabilities.
}
