% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{get.step.sequence}
\alias{get.step.sequence}
\title{Calculate the sequence of dating steps}
\usage{
get.step.sequence(datmin = 0, datmax = 100, stepsize = 25)
}
\arguments{
\item{datmin}{todo}

\item{datmax}{todo}

\item{stepsize}{todo}
}
\value{
sequence
}
\description{
TO DO also i dont want to write documentation
}
