% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_funs.R
\name{scaleweight}
\alias{scaleweight}
\title{Scales the content of the weight columns according to group membership}
\usage{
scaleweight(df, var)
}
\arguments{
\item{df}{a dataframe as returned by datsteps}

\item{var}{the columns of said dataframe that should be used as the group variable}
}
\value{
the same dataframe, with the 'weight'-values scaled along group membership
}
\description{
Requires a dataframe as produces by datsteps(). (Meaning 6 columns in the following order: ID, group, minimum/earliest date, maximum/latest date, weight, 'DAT_Steps')
}
