% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{check.structure}
\alias{check.structure}
\title{Check the Structure to be compatible with datsteps()}
\usage{
check.structure(DAT_df)
}
\arguments{
\item{DAT_df}{A value to check}
}
\value{
TRUE if df can be processen, FALSE if not
}
\description{
Checks if the object passed to datsteps() will work
}
