% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{check.structure}
\alias{check.structure}
\title{Check the Structure to be compatible with datsteps()  (internal)}
\usage{
check.structure(DAT_df)
}
\arguments{
\item{DAT_df}{An object to check}
}
\value{
TRUE if object can be processed by datsteps(), error / FALSE if not
}
\description{
Checks if the object passed to datsteps() will work
}
