% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datsteps.R
\name{datsteps}
\alias{datsteps}
\title{Create 'steps' of dates for each object in a dataframe}
\usage{
datsteps(DAT_df, stepsize = 25, calc = "weight", cumulative = FALSE)
}
\arguments{
\item{DAT_df}{a dataframe with 4 variables: ID, group, minimum date (int/num)
maximum date (int/num), _must_ be in this order, colnames are irrelevant;
each object _must_ be one row.}

\item{stepsize}{defaults to 5. Number of years that should be used as an
interval for creating dating steps.}

\item{calc}{"weight" (default): use the published original
calculation (https://doi.org/10.1017/aap.2021.8) for weights,
"probability": calculate year-wise probability instead (only useful for a stepsize
of 1)}

\item{cumulative}{TRUE: add a column for cumulative probability ("cumul_prob")
(only useful for a stepsize of 1, as the summed probability in larger
stepsizes has no meaning)}
}
\value{
a data.frame with a number of steps for each object as well
as a 'weight' value, that is a quantification of how well the object is
dated (lesser value means object is dated to larger timespans,
i.e. with less confidence)
}
\description{
Requires a dataframe with 4 variables: ID (ideally factor),
group (ideally factor), minimum date (int/numeric) and maximum date
(int/numeric). It's expected that dates BCE are displayed as negative
values while dates CE are positive values. Ignoring this will cause problems
in any case.
}
\examples{
DAT_df_steps <- datsteps(DAT_df[1:100, ], stepsize = 25)
plot(density(DAT_df_steps$DAT_step))


}
