% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{generate.stepsize}
\alias{generate.stepsize}
\title{Generate stepsize}
\usage{
generate.stepsize(DAT_mat)
}
\arguments{
\item{DAT_mat}{todo}
}
\value{
stepsize
}
\description{
todo
}
