% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_funs.R
\name{generate.stepsize}
\alias{generate.stepsize}
\title{Generate stepsize}
\usage{
generate.stepsize(timespans)
}
\arguments{
\item{df}{a dataframe with 4 variable: ID, group, minimum date (int/num) maximum date (int/num)}
}
\value{
stepsize
}
\description{
Requires a dataframe with 4 variables: ID (ideally factor), group (ideally factor),
minimum date (int/numeric) and maximum date (int/numeric).
}
