% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.histogramscale.R
\name{get.histogramscale}
\alias{get.histogramscale}
\title{Scaling Factor for Combined Histogramm Plots}
\usage{
get.histogramscale(DAT_df, bw = "auto")
}
\arguments{
\item{DAT_df}{a dataframe as returned by datsteps}

\item{bw}{the bandwidtg to use. shoudl be stepsize used to compute the dataframe, and if a df as returned by datsteps() is automatically assigned}
}
\value{
the factor with which to scale the density curve to a histogram plot
}
\description{
Requires a dataframe as produced by datsteps(). (Meaning 6 columns in the following order: ID, group, minimum/earliest date, maximum/latest date, weight, 'DAT_Steps')
}
