% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{get.weights}
\alias{get.weights}
\title{Calculate the weights for each dated object}
\usage{
get.weights(DAT_min, DAT_max)
}
\arguments{
\item{DAT_min}{a vector containing the minimum date (int/num) of each object}

\item{DAT_max}{a vector containing the maximum date (int/num) of each object}
}
\value{
the 'weight' value for the datsteps-dataframe, that is a quantification of how well the object is dated (lesser value means object is dated to larger timespans, i.e. with less confidence)
}
\description{
Requires a dataframe with 4 variables: ID (ideally factor), group (ideally factor),
minimum date (int/numeric) and maximum date (int/numeric). It's expected that dates BCE are
displayed as negative values while dates CE are positive values. Ignoring this will cause problems
in any case.
}
