% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_utility.R
\name{get.weights}
\alias{get.weights}
\title{Calculate the weights for each dated object (internal)}
\usage{
get.weights(DAT_min, DAT_max)
}
\arguments{
\item{DAT_min}{a vector containing the minimum date (a number) of each object}

\item{DAT_max}{a vector containing the maximum date (a number) of each object}
}
\value{
the 'weight' value for the datsteps-dataframe, that is a
quantification of how well the object is dated (lesser value means object
is dated to larger timespans, i.e. with less confidence)
}
\description{
Calculates the weights from two vectors of minimum and maximum
dating for each object. Returns a dataframe with the weight in the first
column and FALSE in the second if two rows have the same value in both
min and max dating.
}
