% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleweights.R
\name{scaleweight}
\alias{scaleweight}
\title{Scales the content of a column according to group membership}
\usage{
scaleweight(DAT_df, var = c("all", 2), val = 5)
}
\arguments{
\item{DAT_df}{a dataframe}

\item{var}{the index of the column that should be used as the group variable, OR "all"
(note: all non-numeric values will result in the weight being scaled accross all objects)}

\item{val}{the column that should be scaled (value / sum(values))}
}
\value{
the same dataframe, with scaled values in the specifies column
}
\description{
Requires a dataframe with one variable and one value column.
}
