% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datplot_funs.R
\name{switch.dating}
\alias{switch.dating}
\title{Switch values where dating is in wrong order}
\usage{
switch.dating(df, DAT_err)
}
\arguments{
\item{df}{a dataframe with 4 variable: ID, group, minimum date (int/num) maximum date (int/num)}

\item{DAT_err}{a vector containing the dates in wrong order}
}
\value{
corrected df
}
\description{
Requires a dataframe with 4 variables: ID (ideally factor), group (ideally factor),
minimum date (int/numeric) and maximum date (int/numeric).
}
